New low-level processing and cleanup
