+++ .travis.yml
@@ -1,0 +1,16 @@
+language: python
+dist: xenial
+python:
+- '3.6'
+- '3.7'
+install:
+- pip install -r requirements.txt
+script:
+- python setup.py test
+deploy:
+  provider: pypi
+  user: jelleaalbers
+  password:
+    secure: sZFPmLKtCKvyN/HrPk2gTh8PEx28g9ksv4aBi15WGg14PUWFg53v4gEMH/DrhBc3MDrIA3ib/smMx0airn+XMBDQixwzHg1y+X1KsljOQzfk5MSC+Uv+wB+GkJW2802r00aKL851X0VU2XdjZIxTDd1DcS2QwoQjKSknRrsOtXUUAipH7JJtwKV3WcLGBUn51yygeuFHMOmtca3zlDAFEdk4W29IyAnVBj38PT8ucEWj64OIWSVJ1xf3WWdamFpNz+z6JopHUygXgyIuj8maAeSISWf8n9i2GuX8MzyumFuExklqEjcn1gcKVjmqxWdAHwvf6ul3sfjUsTbXXoPxPx1Oiz8r09H3p8lDCDwexOoh1UrQbsRGCj9oQH8e3gp8deHyosXFmIGm1DeW9efG/S/abDLQ6L9e7S5owYYFw2Muf3blJ2c6nS43k4drsS3ioP1wBsV/Z7HRI1f9NTsKpH4JwAsvIi7cr4BlTyjNip05Narq6eOp8XHSTbt/u7gfN1gfZdxwhPryRAjI+M2ck6sk6eq6nPA5DXd1B4Ekti7LwVaIBgcPCgbn2R0xaL+FrWU++6IYwhNqmnHOq9Hv3iLenGF+Mp3aTJCiyw8hvAbxqbe2vUSO6VwBQW8W8ZM3H8diU5+UaWlj+bZ0nyjxlB/arQle/6Y/WX8CHc9vBYg=
+  on:
+    tags: true
