+++ setup.py
@@ -16,9 +16,11 @@ setuptools.setup(name='straxen',
                  description='Streaming analysis for XENON',
                  author='Straxen contributors, the XENON collaboration',
                  url='https://github.com/XENONnT/straxen',
-                 install_requires=requires,
                  long_description=readme + '\n\n' + history,
                  long_description_content_type="text/markdown",
+                 setup_requires=['pytest-runner'],
+                 install_requires=requires,
+                 tests_require=requires + ['pytest'],
                  python_requires=">=3.6",
                  extras_require={
                      'docs': ['sphinx',
