+++ straxen/plugins/__init__.py
@@ -1,1 +1,20 @@
-from . import cut_plugins, pax_interface, plugins
+from . import daqreader
+from .daqreader import *
+
+from . import peak_processing
+from .peak_processing import *
+
+from . import pax_interface
+from .pax_interface import *
+
+from . import pulse_processing
+from .pulse_processing import *
+
+from . import peak_processing
+from .peak_processing import *
+
+from . import event_processing
+from .event_processing import *
+
+from . import cuts
+from .cuts import *
