--- straxen/utils.py
@@ -1,7 +1,0 @@
-import collections
-
-import strax
-export, __all__ = strax.exporter()
-
-
-
